🚚 Rename: 업로드 페이지 파일 구조 변경 (#227)
